+++ src/Extension/Core/Shape/DefaultShapeSet.php
@@ -1,7 +1,9 @@
 <?php
 
-namespace PhpTui\Tui\Shape;
+namespace PhpTui\Tui\Extension\Core\Shape;
 
+use PhpTui\Tui\Extension\Core\Shape\CirclePainter;
+use PhpTui\Tui\Extension\Core\Shape\ClosurePainter;
 use PhpTui\Tui\Model\Canvas\ShapeSet;
 
 class DefaultShapeSet implements ShapeSet
