+++ src/Extension/Core/Shape/PointsPainter.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace PhpTui\Tui\Shape;
+namespace PhpTui\Tui\Extension\Core\Shape;
 
+use PhpTui\Tui\Extension\Core\Shape\Points;
 use PhpTui\Tui\Model\Canvas\ShapePainter;
 use PhpTui\Tui\Model\Widget\FloatPosition;
 use PhpTui\Tui\Model\Canvas\Painter;
