+++ src/Extension/Core/Shape/RectanglePainter.php
@@ -1,7 +1,9 @@
 <?php
 
-namespace PhpTui\Tui\Shape;
+namespace PhpTui\Tui\Extension\Core\Shape;
 
+use PhpTui\Tui\Extension\Core\Shape\Line;
+use PhpTui\Tui\Extension\Core\Shape\Rectangle;
 use PhpTui\Tui\Model\Canvas\ShapePainter;
 use PhpTui\Tui\Model\Canvas\Painter;
 use PhpTui\Tui\Model\Canvas\Shape;
