+++ src/Widget/ChartRenderer.php
@@ -12,7 +12,7 @@ use PhpTui\Tui\Model\Widget\HorizontalAlignment;
 use PhpTui\Tui\Model\Widget\LineSet;
 use PhpTui\Tui\Model\Widget\Span;
 use PhpTui\Tui\Model\Canvas\CanvasContext;
-use PhpTui\Tui\Shape\Points;
+use PhpTui\Tui\Extension\Core\Shape\Points;
 use PhpTui\Tui\Widget\Chart\ChartLayout;
 use RuntimeException;
