+++ tests/Model/DisplayTest.php
@@ -10,7 +10,7 @@ use PhpTui\Tui\Model\Marker;
 use PhpTui\Tui\Model\Position;
 use PHPUnit\Framework\TestCase;
 use PhpTui\Tui\Widget\Canvas;
-use PhpTui\Tui\Shape\Points;
+use PhpTui\Tui\Extension\Core\Shape\Points;
 use PhpTui\Tui\Widget\Paragraph;
 use PhpTui\Tui\Widget\RawWidget;
