+++ tests/Shape/CircleTest.php
@@ -8,7 +8,7 @@ use PhpTui\Tui\Model\Buffer;
 use PhpTui\Tui\Model\Marker;
 use PhpTui\Tui\Widget\Canvas;
 use PhpTui\Tui\Model\Canvas\CanvasContext;
-use PhpTui\Tui\Shape\Circle;
+use PhpTui\Tui\Extension\Core\Shape\Circle;
 use Generator;
 
 class CircleTest extends ShapeTestCase
