+++ tests/Shape/LineTest.php
@@ -8,7 +8,7 @@ use PhpTui\Tui\Model\Buffer;
 use PhpTui\Tui\Model\Marker;
 use PhpTui\Tui\Widget\Canvas;
 use PhpTui\Tui\Model\Canvas\CanvasContext;
-use PhpTui\Tui\Shape\Line;
+use PhpTui\Tui\Extension\Core\Shape\Line;
 use Generator;
 
 class LineTest extends ShapeTestCase
